Fixes #15442 - use separate settings for client certs
They might differ form the certs used for https.

Also, we now detect the need for certs based on the schema of
the callback url. I've unified the callback_url and foreman_url so that
it always goes directly to foreman, not though the proxy. It saves
us one request to proxy + we would need to update the trusted hosts
in proxy to trust the requests from local host. It's similar to
how the Puppet's node.rb works, where the node.rb itself does the
call to Foreman, not the proxy.

I've updated the settings.yml.example to start with http by default,
for development purposes, not expecting the devels to have the
certs in place.

As part of making the new authentication to work, the console_auth
option was properly moved to core settings.

To avoid being able to use any valid certificate for accessing the core
service, we now check the serial of the cert to match the one the core
is configured with.