Fixes #26374 - use proxy code for authorization
Also, I've made the core authorization code a bit more defensive:
I've noticed that Puma is setting 'https' in `ENV['HTTPS']`: if we don't
check that we should use https, we would be letting the requests though
even if we shouldn't. Luckily, we're not running Puma just yet, but it's
good to be prepared.