Fixes #26969 - Allow using remote core even if core is available
The behavior can be configured using the external_core key in the
settings.

If unset, the behavior will remain the same, eg. use internal
core if available, fallback to external core otherwise. Setting
it to true will force the proxy to always use an external core.
Setting it to false will make the proxy try to always use the
internal core and disable the feature if it is not available.