Performance related tweaks (#117)
* Fixes #36465 - Get execution plan info effectively

Before this change when getting execution plan counts or status
information we used to load them all and then either count them or pluck
the bits we needed. With this patch, we use specialised APIs which give
us what we need directly.

* Refs #36465 - Drop a copy of data from callback action

The callback action doesn't need to keep the data once it sends it out.

* Refs #36465 - Drop a copy of data from batch callback action

The batch callback action doesn't need to keep the data once it sends
it out.

* Refs #36465 - Drop output chunks one they are stored in output

Output chunks don't need to be stored once we aggregate them in action's
output.

* Refs #36465 - Make batch action finish early

Before it stayed running until all of its sub-tasks finished, even
though it didn't need to.